requirements.txt
django-discover-runner
